+++ config/test.exs
@@ -1,0 +1,5 @@
+use Mix.Config
+
+config :caylir_test, Caylir.TestHelpers.Graph,
+  host: "localhost",
+  port: 64210