+++ test/helpers/graph.exs
@@ -1,0 +1,3 @@
+defmodule Caylir.TestHelpers.Graph do
+  use Caylir.Graph, otp_app: :caylir_test
+end