+++ test/test_helper.exs
@@ -1,1 +1,10 @@
+Code.require_file("helpers/graph.exs", __DIR__)
+
+
+Supervisor.start_link(
+  [ Caylir.TestHelpers.Graph.child_spec ],
+  strategy: :one_for_one
+)
+
+
 ExUnit.start()