+++ CHANGELOG.md
@@ -6,6 +6,9 @@
   - Allows retrieving the shape of a query
   - Uses `:hackney` instead of `:inets`
 
+- Backwards incompatible changes
+  - Configuration is done using config files
+
 ## v0.1.0 (2015-06-04)
 
 - Initial Release